feat: place관련 타입 선언
