+++ src/types/bookmarks.ts
@@ -1,5 +1,5 @@
 import type { ApiSuccessResponse } from '@/types/api';
-import type { PlaceList } from '@/types/video';
+import type { PlaceList } from '@/types/place';
 
 export interface BookmarkResponse extends ApiSuccessResponse {
   data: {
@@ -25,4 +25,20 @@ export interface CreateBookmarkRequest {
 
 export interface CreateBookmarkResponse extends ApiSuccessResponse {
   data: number;
+}
+
+export type AddBookmarkResponse = CreateBookmarkResponse;
+
+export interface AddBookmarkRequest {
+  latitude: number;
+  longitude: number;
+  roadAddress: string;
+  placeName: string;
+  phoneNumber: string;
+  type: 'LARGE_MART';
+}
+
+export interface AddBookmarkProps {
+  bookmarkId: number;
+  data: AddBookmarkRequest;
 }
